rename(web): profile-cion을 mypage 슬라이스로 이동
